% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovax.R
\name{anovax}
\alias{anovax}
\alias{anovax.lmerMod}
\alias{anovax.gls}
\alias{anovax.lm}
\alias{print.anovax}
\title{anova like function}
\usage{
anovax(object, ..., test = "kr", cl = NULL)

\method{anovax}{lmerMod}(object, ..., test = "kr", cl = NULL)

\method{anovax}{gls}(object, ..., test = "pb", cl = NULL)

\method{anovax}{lm}(object, ..., test = "pb", cl = NULL)

\method{print}{anovax}(x, ...)
}
\arguments{
\item{object}{A model object object}

\item{...}{further arguments}

\item{test}{A character string}

\item{cl}{number of cluster (for test="pf" and "pbf")}

\item{x}{anovax object}
}
\description{
anova like function

print anovax object
}
\examples{
fm1 <- lmer(sugpct ~ block + sow + harvest + (1|block:harvest), data=beets)
fm0 <- update(fm1, .~. - sow)
anovax(fm1, ~.-harvest, test="KR")
anovax(fm1, ~.-harvest, test="SAT")
anovax(fm1, ~.-harvest, test="PB", cl=1)

anovax(fm1, test="KR")
anovax(fm1, test="SAT")
anovax(fm1, test="PB", cl=1)
anovax(fm1, test="PBF", cl=1)
}
\author{
Søren Højsgaard
}
