% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovax.R
\name{anovax_list}
\alias{anovax_list}
\title{Various different tests for model comparison}
\usage{
anovax_list(
  object,
  object2,
  test = c("x2", "kr", "sat", "pb"),
  control = list(nsim = 1000)
)
}
\arguments{
\item{object}{Model object}

\item{object2}{Model object or equivalent way of specifying a submodel of lmm1}

\item{test}{A vector with the various test types.}

\item{control}{A list controlling the model comparions.}
}
\value{
Dataframe with results of the various tests
}
\description{
Various different tests for model comparison
}
\author{
Søren Højsgaard
}
