% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbkr_utilities.R
\name{as.data.frame.PBmodcomp}
\alias{as.data.frame.PBmodcomp}
\alias{as.data.frame.summary_PBmodcomp}
\title{Coerce PBmodcomp objects to data frames}
\usage{
\method{as.data.frame}{PBmodcomp}(x, ...)

\method{as.data.frame}{summary_PBmodcomp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{PBmodcomp} or \code{summary_PBmodcomp}.}

\item{...}{Further arguments (currently ignored).}
}
\value{
A \code{data.frame} containing the test results with attributes preserved.
}
\description{
Convert the test result table of a \code{PBmodcomp} or \code{summary.PBmodcomp} object
into a standard data frame for easy inspection or export.
}
\details{
These methods extract the internal \code{test} component (a data frame of test statistics and p-values)
from the object and return it as a standard data frame. The original object's attributes are preserved
as attributes on the returned data frame.

This is useful for exporting or manipulating bootstrap test results in a tidy form.
}
\examples{
\dontrun{
  sleepstudy$Days2 <- sleepstudy$Days^2
  lmer_fit1 <- lme4::lmer(Reaction ~ Days + Days2 + (1 | Subject), data = sleepstudy, REML=FALSE)
  lmer_fit0 <- update(lmer_fit1, . ~ . - Days2)

  res <- pb2_modcomp(lmer_fit1, lmer_fit0, nsim = 200)
  df  <- as.data.frame(res)
  head(df)

  s <- summary(res)
  as.data.frame(s)
}

## @seealso \code{\link{pb2_modcomp}}, \code{\link{summary.PBmodcomp}}

}
