% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovax.R
\name{comodex}
\alias{comodex}
\alias{comodex.lmerMod}
\alias{comodex.gls}
\alias{comodex.glmerMod}
\alias{comodex.lm}
\title{Model comparison}
\usage{
comodex(
  largeModel,
  smallModel,
  test = "x2",
  control = list(),
  details = 0,
  ...
)

\method{comodex}{lmerMod}(
  largeModel,
  smallModel,
  test = "x2",
  control = list(),
  details = 0,
  ...
)

\method{comodex}{gls}(
  largeModel,
  smallModel,
  test = "x2",
  control = list(),
  details = 0,
  ...
)

\method{comodex}{glmerMod}(
  largeModel,
  smallModel,
  test = "x2",
  control = list(),
  details = 0,
  ...
)

\method{comodex}{lm}(
  largeModel,
  smallModel,
  test = "x2",
  control = list(),
  details = 0,
  ...
)
}
\arguments{
\item{largeModel}{An \code{lmer} model}

\item{smallModel}{An \code{lmer} model or a restriction matrix}

\item{test}{A character string}

\item{control}{A list controlling the model comparions.}

\item{details}{should details be printed}

\item{...}{Additional arguments to be passed on to other methods}
}
\description{
Wrapper for functions KRmodcomp, SATmodcomp, PBmodcomp, PBFmodcomp
}
\examples{
(lmer0 <- lmer(Reaction ~ (Days|Subject), sleepstudy))
(lmer1 <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy))
(lmer2 <- lmer(Reaction ~ Days + I(Days^2) + (Days|Subject), sleepstudy))

anovax(lmer2, lmer1)
anovax(lmer2, lmer1, test="sat")
anovax(lmer2, .~. - I(Days^2))
anovax(lmer2, lmer1, test="PB", control=list(nsim=50, cl=1))
anovax(lmer2, lmer1, test="PBF", control=list(nsim=50, cl=1))
}
\author{
Søren Højsgaard
}
