% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAT-modcomp.R
\name{compute_auxiliary}
\alias{compute_auxiliary}
\title{Compute_auxiliary quantities needed for the Satterthwaite
approximation.}
\usage{
compute_auxiliary(model, tol = 1e-06)
}
\arguments{
\item{model}{A linear mixed model object}

\item{tol}{A tolerance}
}
\value{
A list
}
\description{
Computes vcov of variance parameters (theta, sigma), jacobian of
each variance parameter etc.
}
\details{
The code is greatly inspired by code from the lmerTest
    package.
}
\author{
Søren Højsgaard
}
\keyword{internal}
