% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_modcomp.R
\name{fanova}
\alias{fanova}
\alias{fanova.lmerMod}
\alias{fanova.gls}
\alias{fanova.lm}
\alias{print.fanova}
\title{anova like function}
\usage{
fanova(object, ..., test = "kr", cl = NULL)

\method{fanova}{lmerMod}(object, ..., test = "kr", cl = NULL)

\method{fanova}{gls}(object, ..., test = "pb", cl = NULL)

\method{fanova}{lm}(object, ..., test = "pb", cl = NULL)

\method{print}{fanova}(x, ...)
}
\arguments{
\item{object}{A model object object}

\item{...}{further arguments}

\item{test}{A character string}

\item{cl}{number of cluster (for test="pf" and "pbf")}
}
\description{
anova like function
}
\examples{
fm1 <- lmer(sugpct ~ block + sow + harvest + (1|block:harvest), data=beets)
fm0 <- update(fm1, .~. - sow)
fanova(fm1, ~.-harvest, test="KR")
fanova(fm1, ~.-harvest, test="SAT")
fanova(fm1, ~.-harvest, test="PB", cl=1)

fanova(fm1, test="KR")
fanova(fm1, test="SAT")
fanova(fm1, test="PB", cl=1)
fanova(fm1, test="PBF", cl=1)
}
\author{
Søren Højsgaard
}
