% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_modcomp.R
\name{fmodcomp}
\alias{fmodcomp}
\alias{fmodcomp.lmerMod}
\alias{fmodcomp.gls}
\alias{fmodcomp.lm}
\title{Model comparison}
\usage{
fmodcomp(largeModel, smallModel, test = "kr", details = 0, ...)

\method{fmodcomp}{lmerMod}(largeModel, smallModel, test = "kr", details = 0, ...)

\method{fmodcomp}{gls}(largeModel, smallModel, test = "pb", details = 0, ...)

\method{fmodcomp}{lm}(largeModel, smallModel, test = "pb", details = 0, ...)
}
\arguments{
\item{largeModel}{An \code{lmer} model}

\item{smallModel}{An \code{lmer} model or a restriction matrix}

\item{test}{A character string}

\item{details}{should details be printed}

\item{...}{Additional arguments to be passed on to other methods}

\item{cl}{number of clusters}
}
\description{
Wrapper for functions KRmodcomp, SATmodcomp, PBmodcomp, PBFmodcomp
}
\examples{
(fm0 <- lmer(Reaction ~ (Days|Subject), sleepstudy))
(fm1 <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy))
(fm2 <- lmer(Reaction ~ Days + I(Days^2) + (Days|Subject), sleepstudy))

fmodcomp(fm2, fm1)
fmodcomp(fm2, fm1, "sat")
fmodcomp(fm2, .~. - I(Days^2))
fmodcomp(fm2, fm1, "PB", nsim=1000, cl=1)
fmodcomp(fm2, fm1, "PBF", nsim=1000, cl=1)
}
\author{
Søren Højsgaard
}
