% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_y.r
\name{get_vary}
\alias{get_vary}
\title{Get variance covariance matrix for lmerMod and gls}
\usage{
get_vary(object)
}
\arguments{
\item{object}{An lmerMod or gls object}
}
\value{
A sparse matrix
}
\description{
Get variance covariance matrix for lmerMod and gls
}
\author{
Søren Højsgaard
}
