% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gls.R
\name{glance.gls}
\alias{glance.gls}
\title{Glance method for nlme::gls objects}
\usage{
glance.gls(x, ...)
}
\arguments{
\item{x}{A fitted model of class "gls" (from nlme).}

\item{...}{Unused (for S3 compatibility).}
}
\value{
one-row data.frame with overall fit statistics.
}
\description{
Glance method for nlme::gls objects
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
