% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{gls_cov_matrix}
\alias{gls_cov_matrix}
\title{Returns covariance matrix for gls model}
\usage{
gls_cov_matrix(model)
}
\arguments{
\item{model}{A gls object}
}
\value{
A matrix
}
\description{
Returns covariance matrix for gls model
}
\author{
Søren Højsgaard
}
