% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{gls_varcorr}
\alias{gls_varcorr}
\title{Variance and correlation parameters from gls object}
\usage{
gls_varcorr(object)
}
\arguments{
\item{object}{A gls object}
}
\description{
Variance and correlation parameters from gls object
}
\author{
Søren Højsgaard
}
