% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-pbkrtest.R
\name{internal}
\alias{internal}
\alias{\%>\%}
\alias{print.PBmodcomp}
\alias{print.summaryPB}
\alias{summary.PBmodcomp}
\alias{plot.PBmodcomp}
\alias{summary.KRmodcomp}
\alias{print.KRmodcomp}
\alias{KRmodcomp_init}
\alias{KRmodcomp_init.lmerMod}
\alias{KRmodcomp_init.mer}
\alias{as.data.frame.XXmodcomp}
\title{Internal functions for the pbkrtest package}
\description{
These functions are not intended to be called directly.
}
