\name{pbkrtest-internal}

\alias{restrictionMatrix2model}
\alias{restrictionMatrix2model.lm}
\alias{restrictionMatrix2model.mer}
\alias{restrictionMatrix2model.lmerMod}

\alias{print.PBmodcomp}
\alias{print.summaryPB}
\alias{summary.PBmodcomp}
\alias{plot.PBmodcomp}

\alias{summary.KRmodcomp}
\alias{print.KRmodcomp}


\alias{as.data.frame.XXmodcomp} %% This must go!!

\title{Internal functions in the pbkrtest package}
\description{
  Internal functions called by other functions.
}
\keyword{internal}
