% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{refit.glm}
\alias{refit.glm}
\title{Refit linear model / generalized linear model to new response}
\usage{
\method{refit}{glm}(object, newresp, ...)
}
\arguments{
\item{object}{A glm object}

\item{newresp}{A new response variable}

\item{...}{Additional arguments, currently unused}
}
\value{
A refitted model object
}
\description{
Refit linear model / generalized linear model to new response
}
\details{
This i a recent addition and is not thoroughly tested.
}
\author{
Søren Højsgaard
}
