% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{refit_gls}
\alias{refit_gls}
\title{Refit a gls object to new response variable}
\usage{
refit_gls(object, newresp)
}
\arguments{
\item{object}{A gls object}

\item{newresp}{A numeric vector}
}
\value{
A gls object
}
\description{
Refit a gls object to new response variable
}
\author{
Søren Højsgaard
}
