% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{refit_lme}
\alias{refit_lme}
\title{Refit a lme object to new response variable}
\usage{
refit_lme(object, newresp)
}
\arguments{
\item{object}{A lme object}

\item{newresp}{A numeric vector}
}
\value{
A gls object
}
\description{
Refit a lme object to new response variable
}
\author{
Søren Højsgaard
}
