% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_lme4_extension.R
\name{residuals0}
\alias{residuals0}
\title{Residuals ignoring random effects (population-level residuals)}
\usage{
residuals0(object, ...)
}
\arguments{
\item{object}{A fitted model object. Supported classes are \code{lmerMod} (from \pkg{lme4}),
\code{lme} or \code{gls} (from \pkg{nlme}).}

\item{...}{Not used.}
}
\value{
A numeric vector of residuals.
}
\description{
Computes residuals of the form \eqn{y - X\beta}, i.e., the difference between the observed response
and the fixed-effects prediction. These are also called population-level or marginal residuals.
For models without random effects (e.g., \code{gls}), this is equivalent to the usual residuals.
}
\examples{
if (require(nlme)) {
  fit_lme <- lme(distance ~ age, random = ~1 | Subject, data = Orthodont)
  head(residuals0(fit_lme))

  fit_gls <- gls(distance ~ age, data = Orthodont)
  head(residuals0(fit_gls))
}

if (require(lme4)) {
  fit_lmer <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
  head(residuals0(fit_lmer))
}

}
