% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAT-modcomp.R
\name{sat-modcomp}
\alias{sat-modcomp}
\alias{SATmodcomp}
\alias{SATmodcomp.lmerMod}
\title{F-test and degrees of freedom based on Satterthwaite approximation}
\usage{
SATmodcomp(
  largeModel,
  smallModel,
  betaH = 0,
  details = 0,
  eps = sqrt(.Machine$double.eps)
)

\method{SATmodcomp}{lmerMod}(
  largeModel,
  smallModel,
  betaH = 0,
  details = 0,
  eps = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{largeModel}{An \code{lmer} model}

\item{smallModel}{An \code{lmer} model or a restriction matrix}

\item{betaH}{A number or a vector of the beta of the hypothesis, e.g. L
beta=L betaH. betaH=0 if modelSmall is a model not a restriction matrix.}

\item{details}{If larger than 0 some timing details are printed.}

\item{eps}{A small number.}
}
\description{
An approximate F-test based on the Satterthwaite approach.
}
\note{
This code is greatly inspired by code in the lmerTest package.
}
\examples{


(fm2 <- lmer(Reaction ~ Days + + I(Days^2) + (Days|Subject), sleepstudy))
## Define 2-df contrast - since L has 2 (linearly independent) rows
## the F-test is on 2 (numerator) df:
L2 <- rbind(c(0, 1, 0), # Note: ncol(L) == length(fixef(fm))
            c(0, 0, 1))

SATmodcomp(fm2, L2)
}
\references{
Ulrich Halekoh, Søren Højsgaard (2014)., A Kenward-Roger
    Approximation and Parametric Bootstrap Methods for Tests in Linear Mixed
    Models - The R Package pbkrtest., Journal of Statistical Software,
    58(10), 1-30., \url{http://www.jstatsoft.org/v59/i09/}
}
\seealso{
\code{\link{getKR}}, \code{\link{lmer}}, \code{\link{vcovAdj}},
    \code{\link{PBmodcomp}}
}
\author{
Søren Højsgaard
    \email{sorenh@math.aau.dk}
}
\keyword{inference}
\keyword{models}
