% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.R
\name{simulate.gls}
\alias{simulate.gls}
\title{Simulate data for 'gls' object}
\usage{
\method{simulate}{gls}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{A gls object}

\item{nsim}{Number of simulations}

\item{seed}{Seed for random number generator}

\item{...}{Additional arguments; currently not used.}
}
\value{
A dataframe with nsim columns
}
\description{
Simulate data for 'gls' object
}
\author{
Søren Højsgaard
}
