% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovax.R
\name{anovax}
\alias{anovax}
\alias{anovax.lmerMod}
\alias{anovax.glmerMod}
\alias{anovax.gls}
\alias{anovax.lm}
\alias{print.anovax}
\title{anova like function}
\usage{
anovax(object, ..., test = "x2", control = list(nsim = 1000, cl = NULL))

\method{anovax}{lmerMod}(object, ..., test = "x2", control = list(nsim = 1000, cl = NULL))

\method{anovax}{glmerMod}(object, ..., test = "x2", control = list(nsim = 1000, cl = NULL))

\method{anovax}{gls}(object, ..., test = "x2", control = list(nsim = 1000, cl = NULL))

\method{anovax}{lm}(object, ..., test = "x2", control = list(nsim = 1000, cl = NULL))

\method{print}{anovax}(x, ...)
}
\arguments{
\item{object}{A model object object}

\item{...}{further arguments}

\item{test}{A character string}

\item{control}{A list controling simulations, only relevant for
parametric bootstrapping.}

\item{x}{anovax object}
}
\description{
anova like function

print anovax object
}
\examples{
fm1 <- lmer(sugpct ~ block + sow + harvest + (1|block:harvest), data=beets)
fm0 <- update(fm1, .~. - sow)
anovax(fm1, .~. - harvest, test="KR")
anovax(fm1, .~. - harvest, test="SAT")
anovax(fm1, .~. - harvest, test="PB", control=list(nsim=50, cl=1))

anovax(fm1, test="KR")
anovax(fm1, test="SAT")
anovax(fm1, test="PB", control=list(nsim=50, cl=1))
## anovax(fm1, test="PBF", control=list(nsim=50, cl=1))
}
\author{
Søren Højsgaard
}
