% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovax.R
\name{comodex}
\alias{comodex}
\alias{comodex.lmerMod}
\alias{comodex.gls}
\alias{comodex.glmerMod}
\alias{comodex.lm}
\title{Model comparison}
\usage{
comodex(largeModel, smallModel, test = "kr", details = 0, ...)

\method{comodex}{lmerMod}(largeModel, smallModel, test = "kr", details = 0, ...)

\method{comodex}{gls}(largeModel, smallModel, test = "pb", details = 0, ...)

\method{comodex}{glmerMod}(largeModel, smallModel, test = "pb", details = 0, ...)

\method{comodex}{lm}(largeModel, smallModel, test = "pb", details = 0, ...)
}
\arguments{
\item{largeModel}{An \code{lmer} model}

\item{smallModel}{An \code{lmer} model or a restriction matrix}

\item{test}{A character string}

\item{details}{should details be printed}

\item{...}{Additional arguments to be passed on to other methods}
}
\description{
Wrapper for functions KRmodcomp, SATmodcomp, PBmodcomp, PBFmodcomp
}
\examples{
(fm0 <- lmer(Reaction ~ (Days|Subject), sleepstudy))
(fm1 <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy))
(fm2 <- lmer(Reaction ~ Days + I(Days^2) + (Days|Subject), sleepstudy))

anovax(fm2, fm1)
anovax(fm2, fm1, "sat")
anovax(fm2, .~. - I(Days^2))
anovax(fm2, fm1, "PB", nsim=1000, cl=1)
anovax(fm2, fm1, "PBF", nsim=1000, cl=1)
}
\author{
Søren Højsgaard
}
