% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAT-modcomp.R
\name{get_covbeta}
\alias{get_covbeta}
\title{Compute covariance of fixed effect parameters as a function of variance parameters of a linear mixed model}
\usage{
get_covbeta(varpar, devfun)
}
\arguments{
\item{varpar}{variance parameters; \code{varpar = c(theta, sigma)}.}

\item{devfun}{deviance function as a function of theta only.}
}
\value{
The covariances matrix of the fixed effects at supplied \code{varpar} values.
}
\description{
At the optimum the covariance is available as \code{vcov(lmer-model)}. This function
computes \code{cov(beta)} at non (RE)ML estimates of \code{varpar}.
}
\author{
Rune Haubo B. Christensen. Adapted to pbkrtest by Søren Højsgaard.
}
\keyword{internal}
