\name{pbkrtest-internal}

\alias{print.PBmodcomp}
\alias{print.summaryPB}
\alias{summary.PBmodcomp}
\alias{plot.PBmodcomp}

\alias{summary.KRmodcomp}
\alias{print.KRmodcomp}

\alias{as.data.frame.PBmodcomp}
\alias{as.data.frame.summaryPB}

\alias{\%>\%}
\alias{detectCores}
\alias{makeCluster}
%\alias{}

%\alias{get_ddf_Lb}
%\alias{get_ddf_Lb.lmerMod}
%\alias{ddf_Lb}

\alias{KRmodcomp_init}
\alias{KRmodcomp_init.lmerMod}
\alias{KRmodcomp_init.lme}
\alias{KRmodcomp_init.gls}

\alias{getME.lme}
\alias{getME.gls}

\alias{fixef.gls}

%% \alias{vcovAdj15}
%% \alias{vcovAdj15_internal}
%% \alias{vcovAdj16}
\alias{vcovAdj16_internal}

\alias{KRmodcomp_init.mer}
\alias{as.data.frame.XXmodcomp} %% This must go!!

\title{Internal functions in the pbkrtest package}
\description{
  Internal functions called by other functions.
}
\keyword{internal}
