\name{pbkrtest-internal}

\alias{print.PBmodcomp}
\alias{print.summaryPB}
\alias{summary.PBmodcomp}
\alias{plot.PBmodcomp}

\alias{summary.KRmodcomp}
\alias{print.KRmodcomp}

\alias{get_ddf_Lb}
\alias{get_ddf_Lb.lmerMod}
\alias{ddf_Lb}

\alias{KRmodcomp_init}
\alias{KRmodcomp_init.mer}
\alias{KRmodcomp_init.lmerMod}


\alias{as.data.frame.XXmodcomp} %% This must go!!

\title{Internal functions in the pbkrtest package}
\description{
  Internal functions called by other functions.
}
\keyword{internal}
